// apps/web/src/v1/services/api/trade-core/instance/gen.ts
export interface RIPage {
  /**
   * 状态码
   * @format int32
   */
  code: number;
  /** 承载数据 */
  data?: IPage;
  /** 返回消息 */
  msg: string;
  /** 是否成功 */
  success: boolean;
}
// ...
export interface _10 {
  /**
   * 创建时间
   * @format date-time
   */
  createTime?: string;
  /** 创建者 */
  creator?: string;
  /**
   * 主键ID
   * @format int64
   */
  id?: number;
  /** 利率 */
  interestRate?: number;
  /** 我的份额 */
  myShares?: number;
  /** 快照 */
  snapshot?: string;
  /** 金库总份额 */
  totalShares?: number;
  /** TVL */
  tvl?: number;
  /** 金库名 */
  vaultName?: string;
}
// ...
    export type RequestParams = {};
    export type RequestQuery = {
      /**
       * 主键id
       * @format int64
       */
      accountId: number;
    };
    export type RequestBody = never;
    export type RequestHeaders = {};
    export type ResponseBody = RListAccountCountVO;
// ...
export namespace FollowManage {
// ...
  export namespace GetFollowmanageVaultpageList {
    export type RequestParams = {};
    export type RequestQuery = {
      /**
       * 当前页
       * @format int32
       */
      current?: number;
      /**
       * 我的交易账号
       * @format int64
       */
      myTradeAccountId?: number;
      /**
       * 我的金库
       * @default false
       */
      myVault?: boolean;
      /** 查询参数 */
      searchParam?: string;
      /**
       * 每页的数量
       * @format int32
       */
      size?: number;
    };
    export type RequestBody = never;
    export type RequestHeaders = {};
    export type ResponseBody = RIPage;
  }
}